+++ inc/core/ghost.h
@@ -7,6 +7,7 @@
 #include <memory>
 
 #include "helper/renderer.h"
+#include "state.h"
 
 #include "entity.h"
 
@@ -16,9 +17,10 @@ enum class GhostType { BLINKY, PINKY, INKY, CLYDE };
 class Ghost : public Entity {
 private:
   GhostType type;
+  bool is_at_home;
 
 public:
-  Ghost(const double cx, const double cy, GhostType type);
+  Ghost(const double cx, const double cy, GhostType type, bool is_at_home);
   Ghost(const Ghost &other) = delete;
   Ghost(Ghost &&other) = delete;
   Ghost &operator=(const Ghost &other) = delete;
@@ -29,20 +31,17 @@ public:
   void show(std::shared_ptr<Renderer> renderer) override;
 
   bool can_go(std::shared_ptr<Map> map, const Direction &dir) const;
-
   bool can_change_direction(std::shared_ptr<Map> map) const;
+  void update(std::shared_ptr<Map> map, std::tuple<int, int> pacman_pos);
 
-  void update(std::shared_ptr<Map> map, std::tuple<int, int> Pacman_pos);
+  void chase_pacman(std::shared_ptr<Map> map, std::tuple<int, int> pacman_pos);
 
-  void chase_pacman(std::tuple<int, int> Pacman_pos);
+  void blinky_chase(std::shared_ptr<Map> map, std::tuple<int, int> pacman_pos);
+  void pinky_chase(std::shared_ptr<Map> map, std::tuple<int, int> pacman_pos);
+  void inky_chase(std::shared_ptr<Map> map, std::tuple<int, int> pacman_pos);
+  void clyde_chase(std::shared_ptr<Map> map, std::tuple<int, int> pacman_pos);
 
-  void blinky_chase(std::tuple<int, int> Pacman_pos);
-
-  void pinky_chase(std::tuple<int, int> Pacman_pos);
-
-  void inky_chase(std::tuple<int, int> Pacman_pos);
-
-  void clyde_chase(std::tuple<int, int> Pacman_pos);
+  void move(void);
 };
 
 #endif // __inc_core_ghost_H__